Create ggplot2

Creates a ggplot2 object using the parameters passed in.

createGgplot(
  x,
  pointSize = 4,
  percentageYAxis = FALSE,
  mainTitle = NULL,
  xAxisLabel = NULL,
  yAxisLabel = NULL,
  fixedXAxisMultiple = TRUE,
  fixedYAxisMultiple = TRUE,
  xAxisDateFormat = "%d/%m/%y",
  xAxisBreaks = NULL,
  yAxisBreaks = NULL,
  colours = spcColours(),
  themeOverride = NULL,
  ...
)

# S3 method for ptd_spc_df
plot(
  x,
  pointSize = 4,
  percentageYAxis = FALSE,
  mainTitle = NULL,
  xAxisLabel = NULL,
  yAxisLabel = NULL,
  fixedXAxisMultiple = TRUE,
  fixedYAxisMultiple = TRUE,
  xAxisDateFormat = "%d/%m/%y",
  xAxisBreaks = NULL,
  yAxisBreaks = NULL,
  colours = spcColours(),
  themeOverride = NULL,
  ...
)

Arguments

x
an object created by spc()

pointSize
Specify the plotting point size for the ggplot output. Default is 2.5.

percentageYAxis
Specify whether the y axis values are percentages. Accepted values are TRUE for percentage y axis, FALSE for integer y axis. Defaults to FALSE.

mainTitle
Specify a character string value for the ggplot title.

xAxisLabel
Specify a character string value for the x axis title.

yAxisLabel
Specify a character string value for the y axis title.

fixedXAxisMultiple
Specify whether, if producing a faceted spc, x axis should be fixed for all facet plots. Accepted values are TRUE for fixed x axes or FALSE for individual x axes.

fixedYAxisMultiple
Specify whether, if producing a faceted spc, y axis should be fixed for all facet plots. Accepted values are TRUE for fixed y axes or FALSE for individual y axes.

xAxisDateFormat
Specify how dates on the x axis should be displayed. Format should be provided as a character string using 'd m Y' etc syntax.

xAxisBreaks
Specify an interval value for breaks on the x axis. Value should be a character string expressing interval length and type, e.g. "3 months", "7 days".

yAxisBreaks
Specify an interval value for breaks on the y axis. Value should be a numeric vector of length 1, either an integer for integer scales or a decimal value for percentage scales. This option is ignored if faceting is in use.

colours
Specify the colours to use in the plot, use the spcColours() function to change defaults.

themeOverride
Specify a list containing ggplot theme elements which can be used to override the default appearance of the plot.

...
currently ignored

Value

The ggplot2 object