SPC Colours

Produces a list of colours that controls the geoms in the plot

ptd_spc_colours(
  common_cause = "#7B7D7D",
  special_cause_improvement = "#289de0",
  special_cause_neutral = "#361475",
  special_cause_concern = "#fab428",
  value_line = "#7B7D7D",
  mean_line = "#000000",
  lpl = "#7B7D7D",
  upl = "#7B7D7D",
  target = "#de1b1b",
  trajectory = "#361475"
)

Arguments

common_cause, special_cause_improvement, special_cause_neutral, special_cause_concern
the colour of the points

value_line
the colour of the line joining the points

mean_line
the colour of the "mean" (average) line

lpl, upl
the colour the the lower and upper process limit lines

target
the colour of the target line

trajectory
the colour of the trajectory line

Value

a list of colours